chore: users management refinement
